airflow-backend/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
Faker==24.7.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.11.0
